+++ Squirrel/Contexts/ApplicationContext.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
-using Squirrel.Models;
+using Squirrel.Entities;
 
 namespace Squirrel.Contexts
 {
