+++ Squirrel/Controllers/AccountController.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using Squirrel.Contexts;
 
 namespace Squirrel.Controllers
 {
@@ -6,6 +7,11 @@ namespace Squirrel.Controllers
     [Route("api/[controller]")]
     public class AccountController : Controller
     {
+        private readonly ApplicationContext context;
 
+        public AccountController(ApplicationContext context)
+        {
+            this.context = context;
+        }
     }
 }
