+++ Squirrel/Entities/Achievement.cs
@@ -1,0 +1,10 @@
+﻿namespace Squirrel.Entities
+{
+    public class Achievement
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public string Description { get; set; }
+        public double Goal { get; set; }
+    }
+}
