+++ Squirrel/Entities/Category.cs
@@ -1,14 +1,7 @@
-﻿namespace Squirrel.Models
+﻿namespace Squirrel.Entities
 {
     public class Category : UserRelatedModel
     {
-        public Category() { }
-        public Category(string name, string color, User user) : base(user)
-        {
-            Name = name;
-            Color = color;
-        }
-
         public int Id { get; set; }
         public string Name { get; set; }
         public string Color { get; set; }
