+++ Squirrel/Entities/UserRelatedModel.cs
@@ -1,0 +1,8 @@
+﻿namespace Squirrel.Entities
+{
+    public class UserRelatedModel
+    {
+        public string UserId { get; set; }
+        public User User { get; set; }
+    }
+}
