+++ Squirrel/Program.cs
@@ -7,9 +7,14 @@ var configuration = builder.Configuration;
 
 services.AddDbContext<ApplicationContext>(options => options.UseSqlite(configuration.GetConnectionString("DefaultConnection")));
 services.AddControllers();
+services.AddSwaggerGen();
 
 var app = builder.Build();
-
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI();
+}
 app.UseHttpsRedirection();
 
 app.UseAuthorization();
